nickname encoding, moved to new file
